Combine local and http into client package, unify tests with table-driven tests
