+++ rpc/client/httpclient.go
@@ -1,5 +1,16 @@
+package client
+
+import (
+	"encoding/json"
+
+	"github.com/pkg/errors"
+	"github.com/tendermint/go-rpc/client"
+	ctypes "github.com/tendermint/tendermint/rpc/core/types"
+	"github.com/tendermint/tendermint/types"
+)
+
 /*
-package http returns a Client implementation that communicates
+HTTP is a Client implementation that communicates
 with a tendermint node over json rpc and websockets.
 
 This is the main implementation you probably want to use in
@@ -7,21 +18,7 @@ production code.  There are other implementations when calling
 the tendermint node in-process (local), or when you want to mock
 out the server for test code (mock).
 */
-package http
-
-import (
-	"encoding/json"
-
-	"github.com/pkg/errors"
-	"github.com/tendermint/go-rpc/client"
-	"github.com/tendermint/tendermint/rpc/client"
-	ctypes "github.com/tendermint/tendermint/rpc/core/types"
-	"github.com/tendermint/tendermint/types"
-)
-
-// Client is a Client implementation that communicates over
-// JSONRPC
-type Client struct {
+type HTTP struct {
 	remote   string
 	endpoint string
 	rpc      *rpcclient.ClientJSONRPC
@@ -30,19 +27,19 @@ type Client struct {
 
 // New takes a remote endpoint in the form tcp://<host>:<port>
 // and the websocket path (which always seems to be "/websocket")
-func New(remote, wsEndpoint string) *Client {
-	return &Client{
+func NewHTTP(remote, wsEndpoint string) *HTTP {
+	return &HTTP{
 		rpc:      rpcclient.NewClientJSONRPC(remote),
 		remote:   remote,
 		endpoint: wsEndpoint,
 	}
 }
 
-func (c *Client) _assertIsClient() client.Client {
+func (c *HTTP) _assertIsClient() Client {
 	return c
 }
 
-func (c *Client) Status() (*ctypes.ResultStatus, error) {
+func (c *HTTP) Status() (*ctypes.ResultStatus, error) {
 	tmResult := new(ctypes.TMResult)
 	_, err := c.rpc.Call("status", []interface{}{}, tmResult)
 	if err != nil {
@@ -52,7 +49,7 @@ func (c *Client) Status() (*ctypes.ResultStatus, error) {
 	return (*tmResult).(*ctypes.ResultStatus), nil
 }
 
-func (c *Client) ABCIInfo() (*ctypes.ResultABCIInfo, error) {
+func (c *HTTP) ABCIInfo() (*ctypes.ResultABCIInfo, error) {
 	tmResult := new(ctypes.TMResult)
 	_, err := c.rpc.Call("abci_info", []interface{}{}, tmResult)
 	if err != nil {
@@ -61,7 +58,7 @@ func (c *Client) ABCIInfo() (*ctypes.ResultABCIInfo, error) {
 	return (*tmResult).(*ctypes.ResultABCIInfo), nil
 }
 
-func (c *Client) ABCIQuery(path string, data []byte, prove bool) (*ctypes.ResultABCIQuery, error) {
+func (c *HTTP) ABCIQuery(path string, data []byte, prove bool) (*ctypes.ResultABCIQuery, error) {
 	tmResult := new(ctypes.TMResult)
 	_, err := c.rpc.Call("abci_query", []interface{}{path, data, prove}, tmResult)
 	if err != nil {
@@ -70,7 +67,7 @@ func (c *Client) ABCIQuery(path string, data []byte, prove bool) (*ctypes.Result
 	return (*tmResult).(*ctypes.ResultABCIQuery), nil
 }
 
-func (c *Client) BroadcastTxCommit(tx types.Tx) (*ctypes.ResultBroadcastTxCommit, error) {
+func (c *HTTP) BroadcastTxCommit(tx types.Tx) (*ctypes.ResultBroadcastTxCommit, error) {
 	tmResult := new(ctypes.TMResult)
 	_, err := c.rpc.Call("broadcast_tx_commit", []interface{}{tx}, tmResult)
 	if err != nil {
@@ -79,15 +76,15 @@ func (c *Client) BroadcastTxCommit(tx types.Tx) (*ctypes.ResultBroadcastTxCommit
 	return (*tmResult).(*ctypes.ResultBroadcastTxCommit), nil
 }
 
-func (c *Client) BroadcastTxAsync(tx types.Tx) (*ctypes.ResultBroadcastTx, error) {
+func (c *HTTP) BroadcastTxAsync(tx types.Tx) (*ctypes.ResultBroadcastTx, error) {
 	return c.broadcastTX("broadcast_tx_async", tx)
 }
 
-func (c *Client) BroadcastTxSync(tx types.Tx) (*ctypes.ResultBroadcastTx, error) {
+func (c *HTTP) BroadcastTxSync(tx types.Tx) (*ctypes.ResultBroadcastTx, error) {
 	return c.broadcastTX("broadcast_tx_sync", tx)
 }
 
-func (c *Client) broadcastTX(route string, tx types.Tx) (*ctypes.ResultBroadcastTx, error) {
+func (c *HTTP) broadcastTX(route string, tx types.Tx) (*ctypes.ResultBroadcastTx, error) {
 	tmResult := new(ctypes.TMResult)
 	_, err := c.rpc.Call(route, []interface{}{tx}, tmResult)
 	if err != nil {
@@ -96,7 +93,7 @@ func (c *Client) broadcastTX(route string, tx types.Tx) (*ctypes.ResultBroadcast
 	return (*tmResult).(*ctypes.ResultBroadcastTx), nil
 }
 
-func (c *Client) NetInfo() (*ctypes.ResultNetInfo, error) {
+func (c *HTTP) NetInfo() (*ctypes.ResultNetInfo, error) {
 	tmResult := new(ctypes.TMResult)
 	_, err := c.rpc.Call("net_info", nil, tmResult)
 	if err != nil {
@@ -105,7 +102,7 @@ func (c *Client) NetInfo() (*ctypes.ResultNetInfo, error) {
 	return (*tmResult).(*ctypes.ResultNetInfo), nil
 }
 
-func (c *Client) DialSeeds(seeds []string) (*ctypes.ResultDialSeeds, error) {
+func (c *HTTP) DialSeeds(seeds []string) (*ctypes.ResultDialSeeds, error) {
 	tmResult := new(ctypes.TMResult)
 	// TODO: is this the correct way to marshall seeds?
 	_, err := c.rpc.Call("dial_seeds", []interface{}{seeds}, tmResult)
@@ -115,7 +112,7 @@ func (c *Client) DialSeeds(seeds []string) (*ctypes.ResultDialSeeds, error) {
 	return (*tmResult).(*ctypes.ResultDialSeeds), nil
 }
 
-func (c *Client) BlockchainInfo(minHeight, maxHeight int) (*ctypes.ResultBlockchainInfo, error) {
+func (c *HTTP) BlockchainInfo(minHeight, maxHeight int) (*ctypes.ResultBlockchainInfo, error) {
 	tmResult := new(ctypes.TMResult)
 	_, err := c.rpc.Call("blockchain", []interface{}{minHeight, maxHeight}, tmResult)
 	if err != nil {
@@ -124,7 +121,7 @@ func (c *Client) BlockchainInfo(minHeight, maxHeight int) (*ctypes.ResultBlockch
 	return (*tmResult).(*ctypes.ResultBlockchainInfo), nil
 }
 
-func (c *Client) Genesis() (*ctypes.ResultGenesis, error) {
+func (c *HTTP) Genesis() (*ctypes.ResultGenesis, error) {
 	tmResult := new(ctypes.TMResult)
 	_, err := c.rpc.Call("genesis", nil, tmResult)
 	if err != nil {
@@ -133,7 +130,7 @@ func (c *Client) Genesis() (*ctypes.ResultGenesis, error) {
 	return (*tmResult).(*ctypes.ResultGenesis), nil
 }
 
-func (c *Client) Block(height int) (*ctypes.ResultBlock, error) {
+func (c *HTTP) Block(height int) (*ctypes.ResultBlock, error) {
 	tmResult := new(ctypes.TMResult)
 	_, err := c.rpc.Call("block", []interface{}{height}, tmResult)
 	if err != nil {
@@ -142,7 +139,7 @@ func (c *Client) Block(height int) (*ctypes.ResultBlock, error) {
 	return (*tmResult).(*ctypes.ResultBlock), nil
 }
 
-func (c *Client) Commit(height int) (*ctypes.ResultCommit, error) {
+func (c *HTTP) Commit(height int) (*ctypes.ResultCommit, error) {
 	tmResult := new(ctypes.TMResult)
 	_, err := c.rpc.Call("commit", []interface{}{height}, tmResult)
 	if err != nil {
@@ -151,7 +148,7 @@ func (c *Client) Commit(height int) (*ctypes.ResultCommit, error) {
 	return (*tmResult).(*ctypes.ResultCommit), nil
 }
 
-func (c *Client) Validators() (*ctypes.ResultValidators, error) {
+func (c *HTTP) Validators() (*ctypes.ResultValidators, error) {
 	tmResult := new(ctypes.TMResult)
 	_, err := c.rpc.Call("validators", nil, tmResult)
 	if err != nil {
@@ -164,7 +161,7 @@ func (c *Client) Validators() (*ctypes.ResultValidators, error) {
 
 // StartWebsocket starts up a websocket and a listener goroutine
 // if already started, do nothing
-func (c *Client) StartWebsocket() error {
+func (c *HTTP) StartWebsocket() error {
 	var err error
 	if c.ws == nil {
 		ws := rpcclient.NewWSClient(c.remote, c.endpoint)
@@ -177,7 +174,7 @@ func (c *Client) StartWebsocket() error {
 }
 
 // StopWebsocket stops the websocket connection
-func (c *Client) StopWebsocket() {
+func (c *HTTP) StopWebsocket() {
 	if c.ws != nil {
 		c.ws.Stop()
 		c.ws = nil
@@ -185,17 +182,17 @@ func (c *Client) StopWebsocket() {
 }
 
 // GetEventChannels returns the results and error channel from the websocket
-func (c *Client) GetEventChannels() (chan json.RawMessage, chan error) {
+func (c *HTTP) GetEventChannels() (chan json.RawMessage, chan error) {
 	if c.ws == nil {
 		return nil, nil
 	}
 	return c.ws.ResultsCh, c.ws.ErrorsCh
 }
 
-func (c *Client) Subscribe(event string) error {
+func (c *HTTP) Subscribe(event string) error {
 	return errors.Wrap(c.ws.Subscribe(event), "Subscribe")
 }
 
-func (c *Client) Unsubscribe(event string) error {
+func (c *HTTP) Unsubscribe(event string) error {
 	return errors.Wrap(c.ws.Unsubscribe(event), "Unsubscribe")
 }
