+++ rpc/client/main_test.go
@@ -1,4 +1,4 @@
-package local_test
+package client_test
 
 import (
 	"os"
@@ -12,11 +12,13 @@ import (
 var node *nm.Node
 
 func TestMain(m *testing.M) {
-	// configure a node, but don't start the server
+	// start a tendermint node (and merkleeyes) in the background to test against
 	app := meapp.NewMerkleEyesApp("", 100)
 	node = rpctest.StartTendermint(app)
-
 	code := m.Run()
+
 	// and shut down proper at the end
+	node.Stop()
+	node.Wait()
 	os.Exit(code)
 }
